Presentation 3/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML Libraries
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
tensorflow==2.15.0
transformers==4.35.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
librosa==0.10.1
soundfile==0.12.1
# facenet-pytorch==2.5.3  # Removed due to installation issues
# huggingface-hub==0.20.1  # Removed due to installation issues

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
neo4j==5.14.0
timescale==0.2.0

# Streaming & Protocols
websockets==12.0
aiortc==1.6.0
kafka-python==2.0.2
pymediainfo==6.1.0

# Security
cryptography==41.0.7
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
webauthn==1.9.0

# Blockchain
hyperledger-fabric-sdk-py==1.0.0
web3==6.11.2

# Utilities
requests==2.31.0
httpx==0.25.2
celery==5.3.4
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
asyncio-mqtt==0.16.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-prometheus==1.12.0

# Image Processing
imageio==2.33.0
scipy==1.11.4
matplotlib==3.8.2

# Audio Processing
pydub==0.25.1
webrtcvad==2.0.10